chore(imagetool): remove manager mixin remnants
